led-matrix-api/pyproject.toml
[project]
name = "led-matrix-api"
version = "0.1.0"
description = "FastAPI service for the LED Matrix controls"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "rgbmatrix",
    "uvicorn>=0.40.0",
]

[tool.uv.sources]
rgbmatrix = { git = "https://github.com/hzeller/rpi-rgb-led-matrix" }
